Move S3 Consumption details as separate component in shareview and make it conditional to type
