+++ frontend/src/modules/Shares/services/getShareObject.js
@@ -56,6 +56,7 @@ export const getShareObject = ({ shareUri, filter }) => ({
           region
           exists
           description
+          datasetType
         }
       }
     }
